--- a/lib/validator.js
+++ b/lib/validator.js
@@ -65,184 +65,133 @@ Validator.prototype.compile = function(schema) {
 			throw new Error("If the schema is an Array, must contain at least one element!");
 		}
 
-		const rules = flatten(schema.map(r => this._processRule(r, null, false)));
-		return this._checkWrapper(rules, true);
+		return this.compileSchemaType(schema);
+	} 
 
-	} else if (schema != null && typeof schema === "object") {
-		const rules = flatten(Object.keys(schema).map(name => this._processRule(schema[name], name, false)));
-		return this._checkWrapper(rules);
-
-	}
-
-	throw new Error("Invalid schema!");
+	return this.compileSchemaObject(schema);
 };
 
-/**
- * Process a rule item & return checker functions
- *
- * @param {Object} rule
- * @param {String} name
- * @param {Boolean} iterate
- */
-Validator.prototype._processRule = function(rule, name, iterate) {
-	const checks = [];
-
-	if (Array.isArray(rule)) {
-		// Compile the multiple schemas
-		checks.push({
-			fn: this.compile(rule),
-			type: "_multi",
-			name: name,
-			schema: rule,
-			iterate: iterate
-		});
-
-		return checks;
+Validator.prototype.compileSchemaObject = function(schemaObject) {
+	if (schemaObject == null || typeof schemaObject !== "object" || Array.isArray(schemaObject)) {
+		throw new Error("Invalid schema!");
 	}
-
-	if (typeof rule === "string") {
-		rule = {
-			type: rule
-		};
-	}
-
-	if (!this.rules[rule.type]) {
-		throw new Error("Invalid '" + rule.type + "' type in validator schema!");
-	}
-
-
-	/**
-	 * !IMPORTANT!: For the functioning of multiRule cases it is important that
-	 * pushing of object and array special rules is done in directly after this
-	 * simple rule. 
-	 * If you which to push other checks, do it before the simple ones or after
-	 * the array special case.
-	 * See the comments in _checkWrapper for further explanation.
-	 */
-	checks.push({
-		fn: this.rules[rule.type],
-		type: rule.type,
-		name: name,
-		schema: rule,
-		iterate: iterate
-	});
-
-	// Nested schema
-	if (rule.type === "object" && rule.props) {
-		// Compile the child schema
-		checks.push({
-			fn: this.compile(rule.props),
-			type: rule.type,
-			name: name,
-			schema: rule,
-			iterate: iterate,
-			secondPart: true //first part is the "primitive" typeof check above
-		});
-	}
-
-	// Array schema
-	if (rule.type === "array" && rule.items) {
-		// Compile the array schema
-		checks.push({
-			fn: this._checkWrapper(this._processRule(rule.items, null, false)),
-			type: rule.type,
-			name: name,
-			schema: rule,
-			iterate: true,
-			secondPart: true //first part is the "primitive" typeof check above
-		});
-	}
-
-	return checks;
-};
-
-/**
- * Create a wrapper function for compiled schema.
- *
- * @param {Array} rules
- * @param {Boolean} isMultipleRules
- */
-Validator.prototype._checkWrapper = function(rules, isMultipleRules) {
+	
 	const self = this;
+	const checks = {};
+	for (let name in schemaObject) {
+		checks[name] = this.compileSchemaType(schemaObject[name]);
+	}
 
-	// Compiled validator function
-	return function(obj, _schema, pathStack) {
-		let errors = [];
-		const count = rules.length;
-		for (let i = 0; i < count; i++) {
-			const check = rules[i];
-			const schema = check.schema;
+	return function(value, _, path, parent) {
+		const errors = [];
+		for (let name in checks) {
+			const propertyPath = (path ? path + "." : "") + name;
+			const res = checks[name](value[name], schemaObject[name], propertyPath, value);
 
-			let value;
-			let stack;
-			if (check.name) {
-				value = obj[check.name];
-				stack = (pathStack ? pathStack + "." : "") + check.name;
-			} else {
-				value = obj;
-				stack = pathStack ? pathStack : "";
+			if (res !== true) {
+				self.handleResult(errors, propertyPath, res);
 			}
-
-			// Check required fields
-			if ((value === undefined || value === null)) {
-				if (check.type === "forbidden")
-					continue;
-
-				if (schema.optional === true)
-					continue;
-
-				if (!Array.isArray(schema)) {
-					self.handleResult(errors, stack, self.makeError("required"));
-					continue;
-				}
-
-			} // else {
-			// Call the checker function
-			if (check.iterate) {
-				let errorInCurrentArray = false;
-				const l = value.length;
-				for (let i = 0; i < l; i++) {
-					let _stack = stack + "[" + i + "]";
-					let res = check.fn.call(self, value[i], schema, _stack, obj);
-					if (res !== true) {
-						errorInCurrentArray = true;
-						self.handleResult(errors, _stack, res);
-					}
-				}
-				/**
-					 * If this is second part of a multiRule array check and the array
-					 * is valid, then the rule is valid.
-					 */
-				if (!errorInCurrentArray && isMultipleRules && check.secondPart) {
-					return true;
-				}
-			} else {
-				let res = check.fn.call(self, value, schema, stack, obj);
-
-				if (isMultipleRules) {
-					if (res === true) {
-						/**
-							 * Object and array checks are divided into two internal checks. In case of a multiRule
-							 * we have to make sure to check both parts. Thus we we continue to also do the second
-							 * check if their is one.
-							 */
-						const nextRule = rules[i + 1];
-						if (nextRule && nextRule.secondPart){
-							continue;
-						}
-						// Jump out after first success and clear previous errors
-						return true;
-					}
-				}
-
-				if (res !== true)
-					self.handleResult(errors, stack, res);
-			}
-			//}
 		}
 
 		return errors.length === 0 ? true : errors;
 	};
+};
+
+Validator.prototype.compileSchemaType = function(schemaType) {
+	if (Array.isArray(schemaType)) {
+		// Multiple rules
+		const self = this;
+		const checks = schemaType.map(r => this.compileSchemaType(r));
+
+		return function(value, _, path, parent) {
+			const errors = [];
+			let validated = false;
+			for (let i = 0; i < checks.length; i++) {
+				const res = checks[i](value, schemaType[i], path, parent);
+
+				if (res !== true) {
+					self.handleResult(errors, path, res);
+				} else {
+					validated = true;
+				}
+			}
+
+			return validated ? true : errors;
+		};
+	}
+
+	return this.compileSchemaRule(schemaType);
+};
+
+Validator.prototype.compileSchemaRule = function(schemaRule) {
+	if (typeof schemaRule === "string") {
+		schemaRule = {
+			type: schemaRule
+		};
+	}
+
+	if (!this.rules[schemaRule.type]) {
+		throw new Error("Invalid '" + schemaRule.type + "' type in validator schema!");
+	}
+
+	const self = this;
+	let checkContents = null;
+	
+	if (schemaRule.type === "object") {
+		if (schemaRule.props) {
+			checkContents = this.compileSchemaObject(schemaRule.props);
+		}
+	} else if (schemaRule.type === "array") {
+		if (schemaRule.items) {
+			checkContents = this.compileSchemaArray(schemaRule.items);
+		}
+	}
+
+	return function(value, _, path, parent) {
+		const res = self.checkRule(value, schemaRule, path, parent);
+		if (res !== true) {
+			return res;
+		}
+
+		if (checkContents) {
+			return checkContents(value, schemaRule, path, parent);
+		}
+
+		return true;
+	}
+};
+
+Validator.prototype.compileSchemaArray = function(schemaType) {
+	const self = this;
+	const checkArrayItem = this.compileSchemaType(schemaType);
+
+	return function (value, _, path, parent) {
+		const errors = [];
+		for (let i = 0; i < value.length; i++) {
+			const arrayItemPath = (path ? path : "") + "[" + i + "]";
+			const res = checkArrayItem(value[i], schemaType, arrayItemPath, value);
+
+			if (res !== true) {
+				self.handleResult(errors, arrayItemPath, res);
+			}
+		}
+
+		return errors.length === 0 ? true : errors;
+	}
+};
+
+Validator.prototype.checkRule = function(value, schemaRule, path, parent) {
+	if (value === undefined || value === null) {
+		if (schemaRule.type === "forbidden")
+			return true;
+
+		if (schemaRule.optional === true)
+			return true;
+
+		return this.makeError("required");
+	}
+
+	return this.rules[schemaRule.type].call(this, value, schemaRule, path, parent);
 };
 
 /**