--- a/lib/validator.js
+++ b/lib/validator.js
@@ -59,65 +59,75 @@ Validator.prototype.validate = function(obj, schema) {
  * @throws {Error} Invalid schema
  */
 Validator.prototype.compile = function(schema) {
+	const self = this;
 	if (Array.isArray(schema)) {
 		// Multiple schemas
 		if (schema.length == 0) {
 			throw new Error("If the schema is an Array, must contain at least one element!");
 		}
 
-		return this.compileSchemaType(schema);
+		var rules = this.compileSchemaType(schema);
+		return function(value) {
+			return self.checkSchemaType(value, rules, undefined, null);
+		}
 	} 
 
-	return this.compileSchemaObject(schema);
+	var rule = this.compileSchemaObject(schema);
+	return function(value) {
+		return self.checkSchemaObject(value, rule, undefined, null);
+	}
 };
 
 Validator.prototype.compileSchemaObject = function(schemaObject) {
 	if (schemaObject === null || typeof schemaObject !== "object" || Array.isArray(schemaObject)) {
 		throw new Error("Invalid schema!");
 	}
-	
-	const self = this;
-	const checks = Object.keys(schemaObject).map(name => ({name: name, fn: this.compileSchemaType(schemaObject[name])}));
 
-	return function(value, _, path, parent) {
-		const errors = [];
-		const checksLength = checks.length;
-		for (let i = 0; i < checksLength; i++) {
-			const check = checks[i];
-			const propertyPath = (path !== undefined ? path + "." : "") + check.name;
-			const res = check.fn(value[check.name], undefined, propertyPath, value);
+	var compiledObject = Object.keys(schemaObject).map(name => {
+		const compiledType = this.compileSchemaType(schemaObject[name]);
+		return {name: name, compiledType: compiledType, isArray: Array.isArray(compiledType)};
+	});
 
-			if (res !== true) {
-				self.handleResult(errors, propertyPath, res);
-			}
+	// Uncomment this line to use uncompiled object validator:
+	// return compiledObject;
+
+	const sourceCode = [];
+	sourceCode.push(`let res;`);
+	sourceCode.push(`let propertyPath;`);
+	sourceCode.push(`const errors = [];`);
+	for (let i = 0; i < compiledObject.length; i++) {
+		const property = compiledObject[i];
+		const name = property.name;
+		sourceCode.push(`propertyPath = (path !== undefined ? path + ".${name}" : "${name}");`);
+		if (Array.isArray(property.compiledType)) {
+			sourceCode.push(`res = this.checkSchemaType(value.${name}, compiledObject[${i}].compiledType, propertyPath, value);`);
+		} else {
+			sourceCode.push(`res = this.checkSchemaRule(value.${name}, compiledObject[${i}].compiledType, propertyPath, value);`);
 		}
+		sourceCode.push(`if (res !== true) {`);
+		sourceCode.push(`\tthis.handleResult(errors, propertyPath, res);`);
+		sourceCode.push(`}`);
+	}
 
-		return errors.length === 0 ? true : errors;
+	sourceCode.push(`return errors.length === 0 ? true : errors;`);
+	
+	var compiledObjectFunction = new Function("value", "compiledObject", "path", "parent", sourceCode.join("\n"));
+
+	var self = this;
+	return function(value, _unused, path, parent) {
+		return compiledObjectFunction.call(self, value, compiledObject, path, parent);
 	};
 };
 
 Validator.prototype.compileSchemaType = function(schemaType) {
 	if (Array.isArray(schemaType)) {
-		// Multiple rules
-		const self = this;
-		const checks = schemaType.map(r => this.compileSchemaType(r));
+		// Multiple rules, flatten to array of compiled SchemaRule
+		const rules = flatten(schemaType.map(r => this.compileSchemaType(r)));
+		if (rules.length == 1) {
+			return rules[0];
+		}
 
-		return function(value, _, path, parent) {
-			const errors = [];
-			const checksLength = checks.length;
-			for (let i = 0; i < checksLength; i++) {
-				const res = checks[i](value, undefined, path, parent);
-
-				if (res !== true) {
-					self.handleResult(errors, path, res);
-				} else {
-					// Jump out after first success and clear previous errors
-					return true;
-				}
-			}
-
-			return errors;
-		};
+		return rules;
 	}
 
 	return this.compileSchemaRule(schemaType);
@@ -130,69 +140,116 @@ Validator.prototype.compileSchemaRule = function(schemaRule) {
 		};
 	}
 
-	const checkRule = this.rules[schemaRule.type];
-	if (!checkRule) {
+	const ruleFunction = this.rules[schemaRule.type];
+	if (!ruleFunction) {
 		throw new Error("Invalid '" + schemaRule.type + "' type in validator schema!");
 	}
 
-	const self = this;
-	let checkContents = null;
-
-	if (schemaRule.type === "object") {
-		if (schemaRule.props) {
-			checkContents = this.compileSchemaObject(schemaRule.props);
-		}
-	} else if (schemaRule.type === "array") {
-		if (schemaRule.items) {
-			checkContents = this.compileSchemaArray(schemaRule.items);
-		}
+	let dataParameter = null;
+	let dataFunction = null;
+	
+	if (schemaRule.type === "object" && schemaRule.props) {
+		dataParameter = this.compileSchemaObject(schemaRule.props);
+		dataFunction = this.checkSchemaObject;
+	} else if (schemaRule.type === "array" && schemaRule.items) {
+		dataParameter = this.compileSchemaType(schemaRule.items);
+		dataFunction = this.checkSchemaArray;
 	}
 
-	return function(value, _, path, parent) {
-		const errors = [];
-		if (value === undefined || value === null) {
-			if (schemaRule.type === "forbidden")
-				return true;
-
-			if (schemaRule.optional === true)
-				return true;
-
-			self.handleResult(errors, path, self.makeError("required"));
-			return errors;
-		}
-
-		const res = checkRule.call(self, value, schemaRule, path, parent);
-		if (res !== true) {
-			self.handleResult(errors, path, res);
-			return errors;
-		}
-
-		if (checkContents !== null) {
-			return checkContents(value, undefined, path, parent);
-		}
-
-		return true;
-	}
+	return {
+		schemaRule: schemaRule,
+		ruleFunction: ruleFunction,
+		dataFunction: dataFunction,
+		dataParameter: dataParameter
+	};
 };
 
-Validator.prototype.compileSchemaArray = function(schemaType) {
-	const self = this;
-	const checkArrayItem = this.compileSchemaType(schemaType);
+Validator.prototype.checkSchemaObject = function(value, compiledObject, path, parent) {
+	
+	if (compiledObject instanceof Function) {
+		return compiledObject(value, undefined, path, parent);
+	}
+	
+	const errors = [];
+	const checksLength = compiledObject.length;
+	for (let i = 0; i < checksLength; i++) {
+		const check = compiledObject[i];
+		const propertyPath = (path !== undefined ? path + "." : "") + check.name;
+		const res = this.checkSchemaType(value[check.name], check.compiledType, propertyPath, value);
 
-	return function (value, _, path, parent) {
+		if (res !== true) {
+			this.handleResult(errors, propertyPath, res);
+		}
+	}
+
+	return errors.length === 0 ? true : errors;
+}
+
+Validator.prototype.checkSchemaType = function(value, compiledType, path, parent) {
+	if (Array.isArray(compiledType)) {
 		const errors = [];
-		const valueLength = value.length;
-		for (let i = 0; i < valueLength; i++) {
-			const arrayItemPath = (path !== undefined ? path : "") + "[" + i + "]";
-			const res = checkArrayItem(value[i], undefined, arrayItemPath, value);
+		const checksLength = compiledType.length;
+		for (let i = 0; i < checksLength; i++) {
+			// Always compiled to list of rules
+			const res = this.checkSchemaRule(value, compiledType[i], path, parent);
 
 			if (res !== true) {
-				self.handleResult(errors, arrayItemPath, res);
+				this.handleResult(errors, path, res);
+			} else {
+				// Jump out after first success and clear previous errors
+				return true;
 			}
 		}
 
-		return errors.length === 0 ? true : errors;
+		return errors;
+	} 
+
+	return this.checkSchemaRule(value, compiledType, path, parent);
+}
+
+Validator.prototype.checkSchemaArray = function(value, compiledArray, path, parent) {
+	const errors = [];
+	const valueLength = value.length;
+	
+	for (let i = 0; i < valueLength; i++) {
+		const itemPath = (path !== undefined ? path : "") + "[" + i + "]";
+		const res = this.checkSchemaType(value[i], compiledArray, itemPath, value);
+
+		if (res !== true) {
+			this.handleResult(errors, itemPath, res);
+		}
 	}
+
+	return errors.length === 0 ? true : errors;
+}
+
+Validator.prototype.checkSchemaRule = function(value, compiledRule, path, parent) {
+	const schemaRule = compiledRule.schemaRule;
+
+	if (value === undefined || value === null) {
+		if (schemaRule.type === "forbidden")
+			return true;
+
+		if (schemaRule.optional === true)
+			return true;
+
+		const errors = [];
+		this.handleResult(errors, path, this.makeError("required"));
+		return errors;
+	}
+
+	const res = compiledRule.ruleFunction.call(this, value, schemaRule, path, parent);
+	if (res !== true) {
+		const errors = [];
+		this.handleResult(errors, path, res);
+		return errors;
+	}
+
+	if (compiledRule.dataFunction !== null) {
+		return compiledRule.dataFunction.call(this, value, compiledRule.dataParameter, path, parent);
+	}
+
+	return true;
 };
 
 /**