--- a/lib/validator.js
+++ b/lib/validator.js
@@ -85,10 +85,10 @@ Validator.prototype.compileSchemaObject = function(schemaObject) {
 
 	var compiledObject = Object.keys(schemaObject).map(name => {
 		const compiledType = this.compileSchemaType(schemaObject[name]);
-		return {name: name, compiledType: compiledType, isArray: Array.isArray(compiledType)};
+		return {name: name, compiledType: compiledType};
 	});
 
-	// Uncomment this line to use uncompiled object validator:
+	// Uncomment this line to use compiled object validator:
 	// return compiledObject;
 
 	const sourceCode = [];
@@ -110,7 +110,7 @@ Validator.prototype.compileSchemaObject = function(schemaObject) {
 	}
 
 	sourceCode.push(`return errors.length === 0 ? true : errors;`);
-	
+
 	var compiledObjectFunction = new Function("value", "compiledObject", "path", "parent", sourceCode.join("\n"));
 
 	var self = this;
@@ -147,7 +147,7 @@ Validator.prototype.compileSchemaRule = function(schemaRule) {
 
 	let dataParameter = null;
 	let dataFunction = null;
-	
+
 	if (schemaRule.type === "object" && schemaRule.props) {
 		dataParameter = this.compileSchemaObject(schemaRule.props);
 		dataFunction = this.checkSchemaObject;
@@ -165,11 +165,10 @@ Validator.prototype.compileSchemaRule = function(schemaRule) {
 };
 
 Validator.prototype.checkSchemaObject = function(value, compiledObject, path, parent) {
-	
 	if (compiledObject instanceof Function) {
 		return compiledObject(value, undefined, path, parent);
 	}
-	
+
 	const errors = [];
 	const checksLength = compiledObject.length;
 	for (let i = 0; i < checksLength; i++) {
@@ -183,7 +182,7 @@ Validator.prototype.checkSchemaObject = function(value, compiledObject, path, pa
 	}
 
 	return errors.length === 0 ? true : errors;
-}
+};
 
 Validator.prototype.checkSchemaType = function(value, compiledType, path, parent) {
 	if (Array.isArray(compiledType)) {
@@ -205,15 +204,15 @@ Validator.prototype.checkSchemaType = function(value, compiledType, path, parent
 	} 
 
 	return this.checkSchemaRule(value, compiledType, path, parent);
-}
+};
 
 Validator.prototype.checkSchemaArray = function(value, compiledArray, path, parent) {
 	const errors = [];
 	const valueLength = value.length;
-	
+
 	for (let i = 0; i < valueLength; i++) {
 		const itemPath = (path !== undefined ? path : "") + "[" + i + "]";
-		const res = this.checkSchemaType(value[i], compiledArray, itemPath, value);
+		const res = this.checkSchemaType(value[i], compiledArray, itemPath, value, parent);
 
 		if (res !== true) {
 			this.handleResult(errors, itemPath, res);
@@ -221,7 +220,7 @@ Validator.prototype.checkSchemaArray = function(value, compiledArray, path, pare
 	}
 
 	return errors.length === 0 ? true : errors;
-}
+};
 
 Validator.prototype.checkSchemaRule = function(value, compiledRule, path, parent) {
 	const schemaRule = compiledRule.schemaRule;