This reorganizes the internals somewhat, compiling to more explicit validators with fewer special cases.

Pseudo BNF for the schema compiler after this PR:

```
RootSchema   ::= SchemaType[] | SchemaObject
SchemaObject ::= map<string, SchemaType>
SchemaType   ::= SchemaType[] | SchemaRule
SchemaRule   ::= string 
	| {type: "object", props?: SchemaObject}
	| {type: "array", items?: SchemaType}
	| {type: string}
```

Compare to the current schema compiler before this PR: 

```
RootSchema   ::= SchemaType[] | SchemaObject
SchemaObject ::= map<string, SchemaType>
SchemaType   ::= SchemaType[] | SchemaRule
SchemaRule   ::= string
	| {type: "object", props?: RootSchema}
	| {type: "array", items?: SchemaType}
	| {type: string}
```

Main difference is the new BNF object props compile from SchemaObject instead of RootSchema. This makes the new compiler stricter, since it throws errors for certain invalid schemas which compiled before.