@andersnm 2018-09-24
Since the PR was created I've added more commits to match the original performance, and then take it one step further (inspired by ajv).

Here are the results from `npm run bench` on my machine before and after this PR

## Before this PR

	compile & validate (#)                       0%      (1,031,134 rps)   (avg: 969ns)
	validate with pre-compiled schema      +285.35%      (3,973,458 rps)   (avg: 251ns)
	validate with wrong obj                  -34.3%        (677,442 rps)   (avg: 1μs)

## After this PR

	compile & validate (#)                       0%        (133,373 rps)   (avg: 7μs)
	validate with pre-compiled schema      +3,724.24%      (5,100,516 rps)   (avg: 196ns)
	validate with wrong obj                +363.96%        (618,795 rps)   (avg: 1μs)

## After this PR (pure JS, uncommented line 92 in validator.js)

	compile & validate (#)                       0%      (1,488,978 rps)   (avg: 671ns)
	validate with pre-compiled schema      +158.51%      (3,849,140 rps)   (avg: 259ns)
	validate with wrong obj                 -59.77%        (599,038 rps)   (avg: 1μs)

@icebob Please let me know if this is intersting. Next I'd like to implement #15 and #30 based on this PR. (I've already implemented #15 in a separate "sanitize" branch)