@icebob 2018-09-25
Great job @andersnm, I will check it, but I need some time.