@icebob 2018-10-20
I'm reviewing the code, but the performance is great. You beat the `ajv`, as well ;) Congratulation!