@icebob 2018-10-25
Released in 0.6.11.
@andersnm so, do you plan to add some other awesome PR? :)